+++ train.py
@@ -6,67 +6,6 @@ from torch.utils.tensorboard import SummaryWriter
 #from models.amino_clust import VQVAE 
 from models.amino_clust_dense import VQVAE 
 
-aminoacid_str = [
-    "A", # Alanine
-    "C", # Cysteine
-    "D", # Aspartic acid
-    "E", # Glutamic acid
-    "F", # Phenylalanine
-    "G", # Glycine
-    "H", # Histidine
-    "I", # Isoleucine
-    "K", # Lysine
-    "L", # Leucine
-    "M", # Methionine
-    "N", # Asparagine
-    "P", # Proline
-    "Q", # Glutamine
-    "R", # Arginine
-    "S", # Serine
-    "T", # Threonine
-    "V", # Valine
-    "W", # Tryptophan
-    "Y" # Tyrosine
-]
-aminoacid_dict = {aminoacid: i+1 for i, aminoacid in enumerate(aminoacid_str)}#.update({"X": 0})  # X for unknown
-reversed_aminoacid_dict = {v: k for k, v in aminoacid_dict.items()}
-
-def prepare_data(config):
-    input_dim = config['model']['input_dim']
-    data_dir = config['base']['train_dir']
-
-    embeddings = []
-    aas = []
-    with open(data_dir, 'r') as f:
-        for line in f:
-            line = line.strip()
-            if not line:
-                continue
-            aa, vec = line.split(':')
-            vec_values = list(map(float, vec.strip().split()))
-            if len(vec_values) != input_dim:
-                raise ValueError(f"Expected {input_dim} values, got {len(vec_values)}")
-            embeddings.append(vec_values)
-            aas.append(aminoacid_dict[aa])
-
-    return embeddings, aas
-
-def data_loader(config):
-    embeddings, aas = prepare_data(config)
-    
-    embeddings = torch.tensor(embeddings, dtype=torch.float32)
-    full_dataset = TensorDataset(embeddings, torch.tensor(aas))
-    total_len = len(full_dataset)
-    train_len = int(0.8 * total_len)
-    val_len = int(0.1 * total_len)
-    test_len = total_len - train_len - val_len
-
-    train_ds, val_ds, test_ds = random_split(full_dataset, [train_len, val_len, test_len])
-    train_loader = DataLoader(train_ds, batch_size=32, shuffle=True)
-    val_loader = DataLoader(val_ds, batch_size=32, shuffle=False)
-    test_loader = DataLoader(test_ds, batch_size=32, shuffle=False)
-
-    return train_loader, val_loader, test_loader
 
 def train_model(config):
     latent_dim = config['model']['latent_dim']
